[BE] refactor : DB 설계 수정
- 도메인듫 하나로 합침

associate with #3
