--- server/src/main/java/com/codesquad/todo/domain/BaseTodo.java
@@ -1,25 +1,0 @@
-package com.codesquad.todo.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-public class BaseTodo {
-
-    @NotNull
-    protected String text;
-
-    @CreatedDate
-    protected LocalDateTime createdAt;
-
-    @LastModifiedDate
-    protected LocalDateTime modifiedAt;
-
-    protected boolean isDeleted;
-}
