+++ server/src/main/java/com/codesquad/todo/domain/Item.java
@@ -3,13 +3,25 @@ package com.codesquad.todo.domain;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import java.time.LocalDateTime;
 
 @Getter
 @Setter
 @NoArgsConstructor
-public class Item extends BaseTodo {
+public class Item {
 
     @Id
     private Long id;
+
+    private String text;
+
+    @CreatedDate
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedAt;
 }
