+++ server/src/main/java/com/codesquad/todo/domain/User.java
@@ -5,6 +5,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
 
+import java.util.List;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,4 +18,6 @@ public class User {
     private String userId;
 
     private String password;
+
+    private List<Item> items;
 }
